# swagger.yaml
openapi: 3.0.0
info:
  title: Admin API
  description: API for managing admin users and chat functionality
  version: 1.0.0
servers:
- url: http://localhost:5764
  description: Local development server
paths:
...
  /courses:
    post:
      summary: "Create a new course"
      description: "Create a course using name and description."
      operationId: "createCourse"
      tags:
      - "Course"
      requestBody:
        description: "Course information for creation"
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
              required:
              - name
              - description
      responses:
        '200':
          description: "Course created successfully"
        '400':
          description: "Only admin can create courses"
        '401':
          description: "Unauthorized"
        '500':
          description: "Error creating course"
...
  /studentDetail:
    get:
      summary: "Get student details"
      description: "Get full details of the student using their ID."
      operationId: "handleStudentDetail"
      tags:
      - "Student"
      requestBody:
      description: "Quiz information for creation"
      content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string 
              required:
              - name
              - description
      responses:
        '200':
          description: "Student details retrieved successfully"
        '400':
          description: "Bad request"
        '404':
          description: "Student not found"
        '500':
          description: "Internal server error"
  /playQuiz/{quizId}:
    post:
      summary: "Start a quiz"
      description: "Play a quiz for the given quizId."
      operationId: "handlePlayQuiz"
      tags:
      - "Student"
      requestBody:
        description: "Quiz information for play"
        content:
          application/json:
            schema:
              type: object
              properties:
                studentId:
                  type: string
                lessonId:
                   type: string
              required:
              - name
              - description
      responses:
        '200':
          description: "Lesson and quizzes fetched and create progress successfully"
        '400':
          description: "Bad request"
        '404':
          description: "Lesson not found"
        '500':
          description: "Internal server error"
  /submitQuiz/{lessonId}:
    post:
      summary: "Submit a quiz"
      description: "Submit answers for a quiz given the lessonId and studentId."
      operationId: "handleSubmitQuiz"
      tags:
      - "Student"
      parameters:
      - in: path
        name: lessonId
        required: true
        description: "ID of the lesson"
        schema:
          type: string
      requestBody:
        description: "Answers and student data for quiz submission"
        content:
          application/json:
            schema:
              type: object
              properties:
                answers:
                  type: array
                  items:
                    type: string
                studentId:
                  type: string
              required:
              - answers
              - studentId
      responses:
        '200':
          description: "Quiz submitted successfully"
        '400':
          description: "Bad request"
        '401':
          description: "Unauthorized"
        '500':
          description: "Error submitting quiz"
  /lessonCompletion:
    get:
      summary: "Get lesson completion progress"
      description: "Get the completion status of a lesson for a student."
      operationId: "getLessonCompletion"
      tags:
      - "Student"
      responses:
        '200':
          description: "Lesson completion details fetched successfully"
        '400':
          description: "Bad request"
        '404':
          description: "Lesson not found"
        '500':
          description: "Internal server error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...